pyproject.toml
[project]
name = "bioscopeai-core"
version = "0.1.0"
description = "Core backend service for BioScopeAI application"
authors = [
    {name = "aleksanderbialka",email = "aleksander.bialka@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi (>=0.120.1,<0.121.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aerich (>=0.9.2,<0.10.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.5,<3.0.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sentry-sdk[fastapi] (>=2.42.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
]

[tool.poetry]
package-mode = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.2,<0.15.0)",
    "vulture (>=2.14,<3.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "pycln (>=2.6.0,<3.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)"
]
test = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "coverage (>=7.11.0,<8.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-env (>=1.2.0,<2.0.0)",
    "pytest-mypy (>=1.0.1,<2.0.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)",
    "pytest-html (>=4.1.1,<5.0.0)",
    "pytest-github-actions-annotate-failures (>=0.3.0,<0.4.0)"
]

# ######################## formatting and linting tools ########################
[tool.pycln]
all = true

[tool.vulture]
make_whitelist = true
min_confidence = 80
paths = ["bioscopeai_core"]
sort_by_size = true
exclude = [".venv", "tests", "migrations", "__init__.py"]
ignore_names = ["cls", "settings_customise_sources"]

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = "3.13"
check_untyped_defs = true
allow_redefinition = false
disallow_any_generics = true
disallow_any_explicit = false
disallow_untyped_calls = true
disallow_untyped_decorators = false
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
explicit_package_bases = true
strict_optional = true
strict_equality = true
local_partial_types = true
warn_no_return = true
warn_unreachable = true
warn_unused_ignores = true
warn_unused_configs = true
warn_redundant_casts = true
disable_error_code = ["misc"]

[[tool.mypy.overrides]]
module = ["tests"]
ignore_errors = true
implicit_reexport = true


# ######################## pytest configuration ########################
[tool.coverage.run]
source = ["bioscopeai_core"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.pytest.ini_options]
minversion = "7.0"
python_files = "tests.py test_*.py *_tests.py"
python_functions = "test_*"
addopts = "--cov=apps --cov-report=html --cov-report=term-missing"


# ################ ################ ruff configuration #################
[tool.ruff]
extend-exclude = ["__pycache__", "*/migrations/*"]
force-exclude = true
namespace-packages = ["bioscopeai_core"]
required-version = ">=0.14.2"
target-version = "py313"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 72

[tool.ruff.lint]
extend-select = [
    "E5",  # pep8 (`E5`)
    "W",  # Warning (`W`)
    "C90",  # mccabe (`C90`)
    "I",  # isort (`I`)
    "N",  # pep8-naming (`N`)
    "UP",  # pyupgrade (`UP`) - eg. UP032: Use f-string instead of format call
    "S",  # flake8-bandit (`S`)
    "BLE",  # flake8-blind-except (`BLE`)
    "B",  # flake8-bugbear (`B`)
    "A",  # flake8-builtins (`A`)
    "COM",  # flake8-commas (`COM`)
    "C4",  # flake8-comprehensions (`C4`)
    "LOG",  # flake8-logging (`LOG`)
    "G",  # flake8-logging-format (`G`) - logging.info("Something happened: %s", var)
    "T20",  # flake8-print (`T20`)
    "PT",  # flake8-pytest-style (`PT`)
    "RUF",
]

ignore = ["E741", "S4", "COM812", "RUF012", "RUF001"]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["S101"]
"migrations/**/*.py" = ["ALL"]
"migrations/*.py" = ["ALL"]

[tool.ruff.lint.flake8-bandit]
check-typed-exception = true

[tool.ruff.lint.isort]
lines-after-imports = 2
order-by-type = false

[tool.ruff.lint.mccabe]
max-complexity = 30

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true
max-doc-length = 88
max-line-length = 100

[tool.aerich]
tortoise_orm = "bioscopeai_core.app.db.init_db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
